mindfulbot/frontend/requirements.txt
Flask
requests
python-dotenv
google-cloud-logging